BlueToothTest20170626

+++ app/src/main/java/com/jedi/wolf_and_hunter/myObj/PlayerInfo.java
@@ -19,7 +19,7 @@ public class PlayerInfo implements Serializable {
     public int teamID;
     public int nowCenterX;
     public int nowCenterY;
-    public int nowFacingAngle;
+    public float nowFacingAngle;
     public boolean judgeingAttack = false;
     public boolean isDead = false;
     public boolean isServer;
